% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btsrp.debias.R
\name{btsrp.debias}
\alias{btsrp.debias}
\title{btsrp.debias}
\usage{
btsrp.debias(data, index, K, func, choice, B, v, alpha)
}
\arguments{
\item{data}{Numeric vector that represents the outcome measures of all the independent observations.}

\item{index}{A list indicates which subgroup or subgroups each subject belongs to. For example, if the nth item in index is the vector c(1,2), the nth subject will be classified into group 1 and group 2 at the same time.}

\item{K}{The number of subgroups.}

\item{func}{A self-defined function that estimates the treatment effect and the standard error using the outcome measures of one subgroup.}

\item{choice}{a vector denotes the range of tuning parameters being chosen from.}

\item{B}{B is the bootstrap times.}

\item{v}{v is the tuning parameter for v-fold cross-validated choice of the tuning parameter.}

\item{alpha}{alpha denotes the level 1 - alpha lower confidence limit.}
}
\value{
\item{lcl}{The estimated 1 - alpha confidence limit for the maximum treatment effect size.}
\item{betamr}{The bias-reduced estimation for the maximum treatment effect size. The details could be found in https://doi.org/10.1080/01621459.2020.1740096}
}
\description{
A function used to deal with the subgroup selection bias based on bootstrap method. You can use this function to estimate the maximum of the parameters.
}
\details{
Dealing with the subgroup selection bias with bootstrap-based method

The order in data and the order in index, which should be corresponding to each other, should both be the sequence of the subject observed.
}
\examples{
# Generate a random sample
set.seed(11); data <- rnorm(300); index <- rep(list(c(1,2),c(1,3),c(2,3)),100); K <- 3; func <- function(x){mean(x)}; choice <- c(1,2,3); B <- 200; v <- 5, alpha <- 0.05

# Print the bias-reduced results
estim <- btsrp.debias(data, index, K, func, choice, B, v, alpha)
print(estim)
summary(estim)

}
