% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdd.debias.R
\name{cfdd.debias}
\alias{cfdd.debias}
\title{cfdd.debias}
\usage{
cfdd.debias(thetas, s, nk, sigm, K, m, B = 2000, cl, cr, M = NULL, R = 200)
}
\arguments{
\item{thetas}{A numeric vector of the estimated treatment effect of the subgroups.}

\item{s}{A numeric vector of the estimated variance of the estimated treatment effect of the subgroups.}

\item{nk}{A vector which denotes the sample size of each subgroup.}

\item{sigm}{sigm is the estimated covariance matrix of the estimated treatment effect of the subgroups.}

\item{K}{The number of subgroups.}

\item{m}{The mth smallest parameter will be estimated.}

\item{B}{The bootstrap times.}

\item{cl}{The range of left bounds being chosen from within the estimation process.}

\item{cr}{The range of right bounds being chosen from within the estimation process.}

\item{M}{The true value of the mth smallest parameter, only for simulation use.}

\item{R}{Tuning parameter.}
}
\value{
\item{beta.hat}{The estimated debiased treatment effect size of interest.}
\item{upper}{The 1 - alpha upper bound of the estimated confidence interval for the treatment effect of interest.}
\item{lower}{The 1 - alpha lower bound of the estimated confidence interval for the treatment effect of interest.}
}
\description{
The function is used to reduce the subgroup selection bias based on confidence-distribution method according to https://doi.org/10.48550/arXiv.2206.11868.
}
\details{
Dealing with the subgroup selection bias with confidence-distribution-based method

you can use this function to estimate any quantile of the parameters of interest compare to the bootstrap-based method.
}
\examples{
# Generate a random sample
set.seed(11); thetas <- rnorm(3); s <- runif(3); nk <- c(200,200,200); sigm <- diag(s); K <- 3; m <- 3; cl <- c(1,2,3); cr <- c(1,2,3)

# Print the bias-reduced results
estim <- cfdd.debias(thetas,s,nk,sigm,K,m,B = 2000,cl,cr,M = NULL,R = 200)
print(estim)
summary(estim)

}
