% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debias.beta_bts.R
\name{debias.beta_bts}
\alias{debias.beta_bts}
\title{An internal function of the de-bias process under the bootstrap-based method}
\usage{
debias.beta_bts(data, index, K, func, r, B, alpha)
}
\arguments{
\item{data}{Numeric vector that represents the outcome measures of all the independent observations.}

\item{index}{A list indicates which subgroup or subgroups each subject belongs to. For example, if the nth item in index is the vector c(1,2), the nth subject will be classified into group 1 and group 2 at the same time.}

\item{K}{The number of the subgroups.}

\item{func}{A self-defined function that estimates the treatment effect and the standard error using the outcome measures of one subgroup.}

\item{B}{The bootstrap times.}

\item{alpha}{Denotes the level 1 - alpha lower confidence limit.}

\item{v}{The tuning parameter for v-fold cross-validated choice of tuning parameter.}
}
\value{
\item{lcl}{The estimated 1 - alpha confidence limit for the maximum treatment effect size.}
\item{betamr}{The bias-reduced estimation for the maximum treatment effect size. The details could be found in https://doi.org/10.1080/01621459.2020.1740096}
}
\description{
An internal function used to estimate the maximum treatment effect of the subgroups and the level 1- alpha confidence limit.
}
\details{
The order in data and the order in index, which should be corresponding to each other, should both be the sequence of the subject observed.
}
\keyword{internal}
