% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estibeta.R
\name{estibeta}
\alias{estibeta}
\title{internal function for debias process with bootstrap-based method}
\usage{
estibeta(data, index, K, func)
}
\arguments{
\item{data}{Numeric vector that represents the outcome measures of all the independent observations.}

\item{index}{A list indicates which subgroup or subgroups each subject belongs to. For example, if the nth item in index is the vector c(1,2), the nth subject will be classified into group 1 and group 2 at the same time.}

\item{K}{The number of the subgroups.}

\item{func}{A self-defined function that estimates the treatment effect and the standard error using the outcome measures of one subgroup.}
}
\value{
\item{beta}{A vector containing the estimated treatment effect of all subgroups.}
\item{sigma}{A vector containing the estimated standard error of the treatment effect of all subgroups.}
}
\description{
This internal function is used to estimate the subgroup treatment effects with the outcome measures and indicators of which subgroup each subject belongs to.
}
\details{
The order in data and the order in index, which should be corresponding to each other, should both be the sequence of the subject observed.
}
\keyword{internal}
