% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debias.beta_cfd.R
\name{newob2}
\alias{newob2}
\title{Internal function for debias process under the confidence-distribution-based method}
\usage{
newob2(the, s, nk)
}
\arguments{
\item{s}{A numeric vector which is the estimated variance of the estimated treatment effect of each subgroup.}

\item{thetas}{A numeric vector which denotes the estimated treatment effect of all the subgroups.}

\item{sigm}{sigm is the estimated covariance matrix of the estimated treatment effect of all the subgroups.}

\item{K}{The number of subgroups.}

\item{m}{The mth smallest parameter will be estimated.}

\item{B}{The bootstrap times.}

\item{cl}{The left bound within the estimation process.}

\item{cr}{The right bound within the estimation process.}

\item{M}{The true value of the mth smallest parameter, only for simulation use.}
}
\value{
\item{esti}{The point estimate for the treatment effect of interest.}
\item{upper}{The 1 - alpha upper bound of the estimated confidence interval for the treatment effect of interest.}
\item{lower}{The 1 - alpha lower bound of the estimated confidence interval for the treatment effect of interest.}
}
\description{
A function used to reduce the subgroup selection bias based on confidence-distribution method and provide a point estimate and a confidence interval.
}
\details{
you can use this function to estimate any quantile of the parameters of interest compare to the bootstrap-based method.
}
\keyword{internal}
