% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btsrp.debias.R
\name{print.btsrpclass}
\alias{print.btsrpclass}
\title{Print method for subgroup selection bias-reduced inference with the bootstrap-based method}
\usage{
\method{print}{btsrpclass}(x, ...)
}
\arguments{
\item{...}{Other arguments.}

\item{object}{Class "btsrpclass" object, obtained by calling \code{\link{btsrp.debias}}.}
}
\value{
\item{}{The lower confidence limit and the estimated debiased beta_max results.}
}
\description{
The print method for subgroup selection bias-reduced inference objects.
}
\examples{
# Generate a random sample
set.seed(11); data <- rnorm(300); index <- rep(list(c(1,2),c(1,3),c(2,3)),100); K <- 3; func <- function(x){mean(x)}; choice <- c(1,2,3); B <- 200; v <- 5, alpha <- 0.05

# Print the bias-reduced results
print(btsrp.debias(data, index, K, func, choice, B, v, alpha))

}
