% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdd.debias.R
\name{print.cfddclass}
\alias{print.cfddclass}
\title{Print method for subgroup selection bias-reduced inference based on confidence-distribution-based method}
\usage{
\method{print}{cfddclass}(x, ...)
}
\arguments{
\item{...}{Other arguments.}

\item{object}{Class "cfddclass" object, obtained by calling \code{\link{cfdd.debias}}.}
}
\value{
\item{}{The estimated debiased beta_max results and the estimated confidence interval.}
}
\description{
The print method for subgroup selection bias-reduced inference objects with confidence-distribution-based method.
}
\examples{
# Generate a random sample
set.seed(11); thetas <- rnorm(3); s <- runif(3); nk <- c(200,200,200); sigm <- diag(s); K <- 3; m <- 3; cl <- c(1,2,3); cr <- c(1,2,3)

# Print the bias-reduced results
print(cfdd.debias(thetas,s,nk,sigm,K,m,B = 2000,cl,cr,M = NULL,R = 200))

}
