% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuningpara_cfd.R
\name{tuning.dist}
\alias{tuning.dist}
\title{tuningpara_cfd}
\usage{
tuning.dist(thetahh, s, nk, sigm, K, m, cl, cr, R, B)
}
\arguments{
\item{s}{The numeric vector which represents the estimated variance of the parameter of each subgroup.}

\item{nk}{nk is a vector which denotes the sample size of each subgroup.}

\item{sigm}{sigm is the estimated covariance matrix of the estimated treatment effect of all the subgroups.}

\item{K}{The number of subgroups.}

\item{m}{The mth smallest parameter will be estimated.}

\item{cl}{The range of left bound within the estimation process being chosen from.}

\item{cr}{The range of right bound within the estimation process being chosen from.}

\item{R}{Tuning parameter.}

\item{B}{The bootstrap times.}

\item{thetah}{The numeric vector which denotes the estimated treatment effect of each subgroup.}
}
\value{
\item{cl}{The selected left bound.}
\item{cr}{The selected right bound.}
}
\description{
A function used to select the best tuning paramters for reducing the subgroup selection bias based on confidence-distribution method.
}
\details{
Internal function for debias process using the confidence-distribution-based method
}
\keyword{internal}
