% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuningpara_bts.R
\name{tuningpara_bts}
\alias{tuningpara_bts}
\title{An internal function for debias process under the bootstrap-based method}
\usage{
tuningpara_bts(data, index, K, func, choice, B, v)
}
\arguments{
\item{data}{a vector represents the outcome measure of all the independent observations.}

\item{index}{a list indicates which subgroup or subgroups subject belongs to.}

\item{K}{K is the number of subgroups.}

\item{func}{func is a self-defined function that estimates the treatment effect and the standard error with the outcome measures of a subgroup.}

\item{choice}{A vector denotes the range of the tuning parameter being chosen from.}

\item{B}{The bootstrap times.}

\item{v}{The tuning parameter for v-fold cross-validated choice of tuning parameter.}
}
\value{
The best selected parameter
}
\description{
The function is used to select the best tuning parameter for dealing with the subgroup selection bias based on bootstrap method.
}
\details{
The order in data and the order in index, which should be corresponding to each other, should both be the sequence of the subject observed.
}
\keyword{internal}
